separates client instantiation
